move some files around, consolidate others
